fix minor bugs while doing test apis

+++ E-learning/src/main/java/com/elearning/Lessons/Dto/LessonDto.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 import java.time.Instant;
 
 @Builder
-public record LessonDto(Integer id, Integer sectionId,Integer index, LessonType type, String title,
+public record LessonDto(Long id, Long sectionId,Integer index, LessonType type, String title,
         LessonState status ,
                                Instant last_published_timeStamp,
                                Instant last_unpublished_timeStamp,
