+++ E-learning/src/main/java/com/elearning/Sections/DTO/SectionDTO.java
@@ -3,5 +3,5 @@ package com.elearning.Sections.DTO;
 import lombok.Builder;
 
 @Builder
-public record SectionDTO(Integer id, Integer index,String title) {
+public record SectionDTO(Long id, Integer index,String title) {
 }
